requirements.txt
numpy
scipy
openai-whisper
ffmpeg
moviepy
numba
pandas
json
csv
io
re
